requirements.txt
pydantic==2.9.2
python-dotenv==1.0.1
fastapi==0.115.5
uvicorn==0.32.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
